requirements.txt
asgiref==3.2.7
Django==3.0.6
django-rest-auth==0.9.5
djangorestframework==3.11.0
Pillow==7.1.2
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
